requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
astroid==3.3.10
attrs==25.3.0
behave==1.2.6
black==25.1.0
cachetools==6.2.1
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
coverage==7.11.0
cryptography==46.0.3
dill==0.4.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.11
flake8==7.0.0
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
load-dotenv==0.1.0
mccabe==0.7.0
model-bakery==1.20.0
mypy_extensions==1.1.0
oauthlib==3.3.1
openai==2.8.0
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.7
pyotp==2.9.0
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.4.1
pytest-cov==6.0.0
pytest-django==4.9.0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tomlkit==0.13.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.8.0
whitenoise==6.11.0
wsproto==1.2.0
 openai-2.8.0